Change how output files are named/generated
Change output folder to `apis/`.

Group versions into 'groups' (e.g. "microsoft.network", etc.).
This mirrors how the files are laid out in the `apis/` dir at
the top level of the repo.

Identify structs by their group+version (AKA `PackageReference`).
This stops some naming conflicts that we had and there are now
an additional 200 files generated.

This does not yet address `import` generation (to be done in a
separate PR for clarity).