Combine multiple types into each output file
Attempt to bundle types for a specific resource into the same
file as that resource. This reduces the number of generated files by
more than 50% (and it should improve with better handling of
AnyOf/OneOf).

The idea is to put each resource into its own file and then put any
types that are _only_ referenced by that resource into the same file.

Types that are referenced by 0 or >1 resources are put into their own
files.