Remove trivial Equal implementations
Removed all Equal methods that match the behaviour of cmp.Equal. Use a
local NodesEqual function that delegates to cmp.Equal with a fixed set
of types for which we allow unexported field comparison.

(I'd have called this `Equal` but that's clobbered by the dot-import
of gomega in the tests.)