Use github.com/pkg/errors to get stack traces from generation errors
Replaces all uses of fmt.Errorf with either errors.New/errors.Errorf
or errors.Wrapf where an existing error was being wrapped with
%w. This means that any errors originating or wrapped in our code will
be annotated with a stack trace that can be included when the error is
logged at the top level, which will make it simpler to track down
problems.

I checked for but didn't find any uses of errors.Is or errors.As to
update to use pkg/errors instead.

As a driveby I also replaced uses of '%s' or '%v' with %q.