Update Type.References to return a TypeNameSet, remove Referees
Returning a set makes the old behaviour simple to implement in the few
places that used it.

Adds a SetUnion to combine two TypeNameSets. I thought about making
this a method to add all of the elements of another set to the
receiver - it would be more efficient but I think it's less clean.