Don't memoise negative checkWithPath results
These can come from a search that is unwinding because it detected a
cycle, rather than one that didn't find a root. In that case we can't
know whether a search from this node with a different path would
succeed.

Added a unit test to explore this problem.

Made the TypeNameSet methods take TypeNames rather than *TypeNames -
this simplified them a bit (at the cost of having to do some extra nil
checks at the edges).